Add CAMMAC and authentication indicator helper functions to libkrb5
This adds k5_unwrap_cammac_svc() to verify and decode a CAMMAC,
and k5_authind_decode() to obtain an array of auth indicator strings
from a CAMMAC authdata.

Also moves the kdc function cammac_check_svcver() to the internal lib
as k5_cammac_check_svcver() to use in the above code.